+++ codes/database/bus.dat
@@ -1,1 +1,2 @@
 (1, Mumbai-pune, 7000, 8000, True, 30, 25)
+(1, Mumbai-pune, 7000, 8000, True, 30, 25)
